requirements.txt
# Core dependencies from project-spec.yml
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.1
google-auth==2.23.4
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.1
plotly==5.17.0
pandas==2.1.3
requests==2.31.0

# Additional useful dependencies
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6